seccion 13 terminada

+++ src/todos/actions/todo-actions.ts
@@ -1,8 +1,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 "use server";
 
+import { getuserSessionServer } from "@/auth";
 import {prisma} from "@/lib/prisma";
 import { Todo } from "@prisma/client";
+import bcrypt from "bcryptjs";
 import { revalidatePath } from "next/cache";
 
 const sleep = (seconds: number)=> {
@@ -12,7 +14,6 @@ const sleep = (seconds: number)=> {
     },seconds * 1000)
   });
 };
-
 
 
 export const toggleTodo = async(id:string, complete:boolean): Promise<Todo>=>{
@@ -38,8 +39,12 @@ export const toggleTodo = async(id:string, complete:boolean): Promise<Todo>=>{
 }
 
 export const addTodo = async(description: string)=>{
+const user = await getuserSessionServer()
+
   try {
-    const todos = await prisma.todo.create({data: {description}});
+    const todos = await prisma.todo.create({
+      data: {description, userId: user!.id},
+    });
 
     revalidatePath("/dashboard/server-todos");
     
@@ -55,13 +60,18 @@ export const addTodo = async(description: string)=>{
 export const deleteCompleted = async (): Promise<
   void | {messaje: string} 
 > => {
+const user = await getuserSessionServer()
+if(!user){
+  return {messaje: "No hay usuario"};
+}
+
   try {
-    const todos = await prisma.todo.findMany({where: {complete: true}});
+    const todos = await prisma.todo.findMany({where: {complete: true, userId: user!.id}});
     if (todos.length === 0) {
       return {messaje: "No hay registros completados"};
     }
 
-    const todo = await prisma.todo.deleteMany({where: {complete: true}});
+    const todo = await prisma.todo.deleteMany({where: {complete: true , userId: user!.id}});
 
     revalidatePath("/dashboard/server-todos");
     return {
@@ -80,15 +90,22 @@ export const getSeed = async (): Promise<{message: string}> => {
   try {
     await prisma.todo.deleteMany();
 
-    await prisma.todo.createMany({
-      data: [
-        {description: "Piedra del alma", complete: true},
-        {description: "Piedra del poder"},
-        {description: "Piedra del tiempo"},
-        {description: "Piedra del espacio"},
-        {description: "Piedra del realidad"},
-      ],
-    });
+    await prisma.user.create({
+        data: {
+          email: "admin@prisma.com",
+          password: bcrypt.hashSync("123456"),
+          roles: ["admin", "client", "super-user"],
+          todos: {
+            create: [
+              {description: "Piedra del alma", complete: true},
+              {description: "Piedra del poder"},
+              {description: "Piedra del tiempo"},
+              {description: "Piedra del espacio"},
+              {description: "Piedra del realidad"},
+            ],
+          },
+        },
+      });
     revalidatePath("/dashboard/server-todos");
 
     return {
